pyproject.toml
[project]
name = "multi-schema-alembic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "typer>=0.16.0",
]
